Human Event Queue

A class for queueing human events. Header file defined here and RcppR6 export handled here. Bound to Human$EventQueue in class Human

HumanEventQ(initQ = 100)

Arguments

initQ
reserve memory for initQ number of events

Methods

firstEvent: return first event in queue as named R list

firstTime: return 'tEvent' slot in first event in queue

rmFirstEventFromQ: remove first event in queue and update queueN

rmTagFromQ(const std::string &tag): remove all events with certain tag from queue and update queueN

get_queueN: return queueN

get_EventQ: return entire EventQ as R list

addEvent2Q(Rcpp::List event): add a list to the EventQ in correct time order and update queueN

Fields

EventQ: std::vector<Rcpp::List>

queueN: integer value of size of queue.