Initialize HumanPop Parameters for HumanPop

make a list of pars for HumanPop

HumanPop.Parameters(nSite, demographics = NULL, ...)

Arguments

nSite
number of sites (either FeedingSite or link{MacroPatch})

demographics
either a list or NULL; if NULL then sitePops is used to generate these parameters, see that function for details on the list structure that is expected.

...
additional named arguments for sitePops

Value

return a list HumanPop_PAR

Examples

HumanPop.Parameters(nSite = 5)

#> $nHumans
#> [1] 41
#> 
#> $tStart
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [39] 0 0 0
#> 
#> $humanIDs
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41
#> 
#> $homeIDs
#>  [1] 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 4 4 4 4 4 4 5 5 5 5 5 5
#> [39] 5 5 5
#> 
#> $bDay
#>  [1] -14421  -6225  -2604  -3436 -14264   -260  -2859  -2753  -4521 -14225
#> [11] -13352 -12644  -2069  -2340 -19096  -3878  -3937 -14532  -1595 -13021
#> [21]   -461  -6093 -15497  -8103  -4748  -5755 -11461  -7563  -1405 -14005
#> [31]  -2745  -5895 -10561  -8608  -9344 -10631  -8768   -875  -1947  -1469
#> [41]  -5392
#> 
#> $biteWeight
#>  [1] 0.095713699 4.346455956 0.533443947 0.991938943 0.587068904 0.900234351
#>  [7] 0.101350057 0.513652551 0.521121069 0.964364365 1.786028221 1.176507510
#> [13] 0.771425885 0.626224050 1.975823910 0.135099666 2.474605658 0.130685188
#> [19] 0.150316922 0.241089996 0.438168718 1.352670740 0.655778431 3.997292067
#> [25] 0.097601701 1.593588690 2.709827213 0.348957088 0.844929685 1.534273094
#> [31] 1.516501345 0.473420142 0.047205864 0.280821777 0.004650806 0.312242653
#> [37] 2.300686987 2.139300961 0.236397300 1.629613976 0.101492580
#>