Generate Parameters for MASH-MBPT MICRO/M-BITES

Generate named list of parameters used throughout MICRO/M-BITES. All arguments have default values which are listed below before the definition.

MBITES.Parameters(enterhouse.a = 9, enterhouse.b = 1, feedHaz.a = 99,
  feedHaz.b = 1, aquaHaz.a = 99, aquaHaz.b = 1, sugarHaz.a = 99,
  sugarHaz.b = 1, swarmHaz.a = 99, swarmHaz.b = 1, F.t = 1,
  B.t = 0.75, R.t = 1.5, L.t = 0.75, O.t = 1, M.t = 1.5, S.t = 0.5,
  F.s = 0.99, B.s = 0.99, L.s = 0.99, O.s = 0.99, M.s = 0.95,
  S.s = 0.95, F.p = 0.95, B.p = 0.98, R.p = 0.98, L.p = 0.8,
  O.p = 0.98, M.p = 0.98, S.p = 0.98, maleM.s = 0.99, maleM.p = 0.95,
  Q = 0.9, reFeed = 0.01, bm.a = 7.5, bm.b = 2.5, rf.a = 60,
  rf.b = 5000, of.a = 5, of.b = 5000, S.a = 20, S.b = 10, S.u = 1/7,
  S.sa = 15, S.sb = 5, sf.a = 7.5, sf.b = 4.5, B.energy = 1/10,
  energyPreG = 0, preGsugar = 0, preGblood = 0, eggP.min = 0.5,
  sns.a = 0.085, sns.b = 100, ttsz.p = 0.5, ttsz.a = 5, ttsz.b = 95,
  ttr.a = 15, ttr.b = 500, maxBatch = 30, E.p = 0.5, Emax = 90,
  Eb = 0.9, eEndm = 180, eEndV = 30, surviveH = 1, probeH = 1,
  surviveprobeH = 1, feedH = 1, surviveL = 1, feedL = 1, SUGAR = TRUE,
  ESTIVATION = FALSE, MATE = TRUE, SENESCE = TRUE, TATTER = TRUE,
  REFEED = TRUE, OVERFEED = TRUE, HISTORY = TRUE, batchSize = "bms",
  eggMatT = "off", InAndOut = matrix(c(c(4, 2, 1, 0, 1), c(2, 1, 1, 0, 1),
  c(2, 1, 1, 0, 2), c(1, 1, 1, 0, 1), c(6, 4, 2, 1, 0)), byrow = FALSE, ncol =
  5), Fwts = rep(1, 5), Rwts = rep(1, 5), Lwts = rep(1, 5),
  Owts = rep(1, 5), Mwts = rep(1, 5), Swts = rep(1, 5))

Arguments

enterhouse.a
= 9, # beta house entry parameters

enterhouse.b
= 1, # beta house entry parameters

feedHaz.a
= 99, # beta feeding site hazard weights

feedHaz.b
= 1, # beta feeding site hazard weights

aquaHaz.a
= 99, # beta aqua site hazard weights

aquaHaz.b
= 1, # beta aqua site hazard weights

sugarHaz.a
= 99, # beta sugar site hazard weights

sugarHaz.b
= 1, # beta sugar site hazard weights

swarmHaz.a
= 99, # beta swarm site hazard weights

swarmHaz.b
= 1, # beta swarm site hazard weights

F.t
= 1, #blood feed search bout timing

B.t
= 0.75, #blood feed attempt bout timing

R.t
= 1.5, #resting bout timing

L.t
= 0.75, #egg laying search bout timing

O.t
= 1, #egg laying attempt bout timing

M.t
= 1.5, #mating bout timing

S.t
= 0.5, #sugar feeding bout timing

F.s
= 0.99,

B.s
= 0.99,

L.s
= 0.99,

O.s
= 0.99,

M.s
= 0.95,

S.s
= 0.95,

F.p
= 0.95,

B.p
= 0.98,

R.p
= 0.98,

L.p
= 0.80,

O.p
= 0.98,

M.p
= 0.98,

S.p
= 0.98,

maleM.s
= .99,

maleM.p
= .95,

Q
= 0.9, #human blood index

reFeed
= 0.01, #probability to refeed post resting bout

bm.a
= 7.5, #shape param for bloodmeal size

bm.b
= 2.5, #shape param for bloodmeal size

rf.a
= 60, #exp param for refeeding as function of bmSize

rf.b
= 5e3, #exp param for refeeding as function of bmSize

of.a
= 5, #exp param for overfeeding as function of bmSize

of.b
= 5e3, #exp param for overfeeding as function of bmSize

S.a
= 20, #exp param for death as function of energy

S.b
= 10, #exp param for death as function of energy

S.u
= 1/7, #energy expenditure for each bout

S.sa
= 15, #exp param for sugar bout as function of energy

S.sb
= 5, #exp param for sugar bout as function of energy

sf.a
= 7.5, #shape param for sugar feed size

sf.b
= 4.5, #shape param for sugar feed size

B.energy
= 1/10, #scaling factor energy from a bloodmeal

energyPreG
= 0, #pre-gonotrophic energy requirement

preGsugar
= 0,

preGblood
= 0,

sns.a
= 0.085, #exp param for senescence

sns.b
= 100, #exp param for senescence

ttsz.p
= 0.5, #zero-inflation for tattering damage

ttsz.a
= 5, #shape param for tattering damage

ttsz.b
= 95, #shape param for tattering damage

ttr.a
= 15, #exp param for tattering survival

ttr.b
= 500, #exp param for tattering survival

maxBatch
= 30, #maximum batch size

E.p
= 0.5, # probability of surviving estivation

Emax
= 90, # onset of the dry season; a day of the calendar year

Eb
= 0.9,

eEndm
= 180, # end of estivation; a day of the calendar year

eEndV
= 30,

surviveH
= 1,

probeH
= 1,

surviveprobeH
= 1,

feedH
= 1,

surviveL
= 1,

feedL
= 1,

SUGAR
= TRUE,

ESTIVATION
= FALSE,

MATE
= TRUE,

SENESCE
= TRUE, #senesce

TATTER
= TRUE, #tattering

REFEED
= TRUE,

OVERFEED
TRUE; simulate overfeeding? See overFeed

HISTORY
TRUE; record history? Needed for cohortBionomics, see historyTrack

batchSize
"bms"; switches for how egg batch size is calculated, should be either "bms" or "norm", see BatchSize.bms or BatchSize.norm

eggMatT
"off"; switches for how egg maturation is calculated, should be either "off" or "norm", see eggMaturationTime.off or eggMaturationTime.norm

InAndOut
matrix of weights for landing spot, see newSpot

Fwts
rep(1,5); blood feeding search bout landing spot weights, see newSpot

Rwts
rep(1,5); resting bout landing spot weights, see newSpot

Lwts
rep(1,5); egg laying search bout landing spot weights, see newSpot

Owts
rep(1,5); egg laying attempt bout landing spot weights, see newSpot

Mwts
rep(1,5); mating bout landing spot weights, see newSpot

Swts
rep(1,5); sugar bout landing spot weights, see newSpot

Value

a named list of parameters

Examples

P = MBITES.PAR()

#> Error in MBITES.PAR(): could not find function "MBITES.PAR"