MICRO Female Mosquito Class Definition

This is a female mosquito class definition for MICRO; it inherits (superclass) from MicroMosquito.

MicroMosquitoFemale

Format

An R6Class generator object

Fields

ID and Time

id: mosquito ID (integer)

bDay: time of emergence (numeric)

tNow: time of current behavioral state (numeric)

tNext: time to next behavioral state change (numeric)

genotype: genotype of mosquito (integer)

State and Location

state: current behavioral state of mosquito (character)

F: Blood Feeding Search Bout

B: Blood Feeding Attempt Bout

R: Post-Prandial Resting Bout

L: Egg Laying Search Bout

O: Egg Laying Attempt Bout

S: Sugar Feeding Attempt Bout

M: Female Mating Bout

E: Estivation Bout

D: Death

stateNew: next behavioral state of mosquito (see above)

inPointSet: class of site (character)

f: feeding site FeedingSite

l: aquatic habitat AquaticSite

s: sugar feeding site SugarSite

m: mating site MatingSite

ix: index of site (integer)

MATURE: mature (logical)

ALIVE: alive or dead? (logical)

Other State Fields

lspot: landing spot (character)

l: Leave the area

r: Reattempt Without Resting

v: Rest on vegetation

w: Rest on the Outside wall of a structure

i: Rest on the Inside wall of a structure

damage: wing tattering (numeric)

energy: energy reserves (numeric)

history: list; see MosquitoFemaleHistory

Egg Production

bmSize: the size of the blood meal, relative to max

batch: female eggs in batch

eggT: the minimum time before eggs are mature

eggP: the mimimum provision for eggs to mature

Maturation & Reproduction

sire: ID of mate

energyPreG: pre-gonotrophic energy requirement

hostID: the id of the host: -1 zoo, 0 null host, otherwise human ID

Pathogen Module Specific Fields

Pathogen: Module specific Pathogen object

PfSI: mosquitoPfSI

Methods

Constructor

new: initialize a new MicroMosquitoFemale object

Arguments:

arg1: something.

Getters & Setters

get_id:

Pointers

get_FemalePopPointer: get pointer to enclosing MicroMosquitoPopFemale

set_FemalePopPointer: set pointer to enclosing MicroMosquitoPopFemale

get_MalePopPointer: get pointer to MicroMosquitoPopMale in same enclosing microsimulation Tile MicroTile

set_MalePopPointer: set pointer to MicroMosquitoPopMale in same enclosing microsimulation Tile MicroTile

get_LandscapePointer: get pointer to Landscape in same enclosing microsimulation Tile MicroTile

set_LandscapePointer: set pointer to Landscape in same enclosing microsimulation Tile MicroTile

get_HumansPointer: get pointer to HumanPop in same enclosing microsimulation Tile MicroTile

set_HumansPointer: set pointer to HumanPop in same enclosing microsimulation Tile MicroTile

get_TilePointer: get pointer to enclosing microsimulation Tile MicroTile

set_TilePointer: set pointer to enclosing microsimulation Tile MicroTile