Initialize PfMOI MODULE

Generate a list of parameters PfMOI_PAR in Human and public methods in Human for PfSI infection model; also defines public methods in MicroMosquitoFemale for PfMOI infection model.

PfMOI.Setup(overwrite = TRUE, MosyMaxI = 1L, Pf_c = 0.15, Pf_b = 0.55,
  DurationPf = 200, LatentPf = 10, FeverPf = 0.3, mnFeverPf = 10,
  vrFeverPf = 0.1, TreatPf = 0.5, mnTreatPf = 3,
  mnChemoprophylaxisPf = 32, PEProtectPf = 0.99, peBlockPf = 1,
  mnPEPf = 270, vrPEPf = 50, GSProtectPf = 1, gsBlockPf = 0.9,
  mnGSPf = 180, vrGSPf = 20, rdtSensPf = 0.9, rdtSpecPf = 0.1,
  lmSensPf = 0.9, lmSpecPf = 0.1)

Arguments

overwrite
overwrite existing methods and fields?

MosyMaxI
maximum number of clonal variants passed in single mosquito to human transmission event (set to -1L for unlimited)

Pf_c
infected human to mosquito transmission efficiency

Pf_b
infected mosquito to human transmission efficiency

DurationPf
duration of infection (assuming clonal variants are cleared independently)

LatentPf
latency (How many days after the infectious bite does the infection start?)

FeverPf
probability of fever

mnFeverPf
mean of timing of fever incident relative to start of PfMOI infection (log-normally distributed, mean on natural scale)

vrFeverPf
standard deviation of timing of fever incident relative to start of PfMOI infection (log-normally distributed, standard deviation on natural scale)

TreatPf
probability of treatment after fever incident

mnTreatPf
average waiting time from fever to treatment (exponentially distributed)

mnChemoprophylaxisPf
constant period of proteection from chemoprophylaxis

PEProtectPf
proportion protected by PE vaccination (probability vaccination successful)

peBlockPf
proportion of infections blocked by PE vaccination

mnPEPf
mean duration of protection from PE vaccination

vrPEPf
standard deviation of protection from PE vaccination

GSProtectPf
proportion protected by GS vaccination (probability vaccination successful)

gsBlockPf
proportion of infections blocked by GS vaccination

mnGSPf
mean duration of protection from GS vaccination

vrGSPf
standard deviation of protection from GS vaccination

rdtSensPf
RDT sensitivity

rdtSpecPf
RDT specificity

lmSensPf
Light Microscopy sensitivity

lmSpecPf
Light Microscopy specificity

Value

Defines a field (list) PfSI_PAR in HumanPop and public methods in Human