Initialize PfSI MODULE

Generate a list of parameters PfSI_PAR in Human and public methods in Human for PfSI infection model; also defines public methods in MicroMosquitoFemale for PfSI infection model.

PfSI.Setup(overwrite = TRUE, Pf_c = 0.15, Pf_b = 0.55, DurationPf = 200,
  LatentPf = 10, FeverPf = 0.3, mnFeverPf = 10, vrFeverPf = 0.1,
  TreatPf = 0.5, mnTreatPf = 1, mnChemoprophylaxisPf = 32,
  PEProtectPf = 0.99, peBlockPf = 1, mnPEPf = 270, vrPEPf = 50,
  GSProtectPf = 1, gsBlockPf = 0.9, mnGSPf = 180, vrGSPf = 20,
  rdtSensPf = 0.9, rdtSpecPf = 0.1, lmSensPf = 0.9, lmSpecPf = 0.1)

Arguments

overwrite
overwrite existing methods and fields?

Pf_c
0.15; transmission efficiency: infected human to mosquito

Pf_b
0.55; transmission efficiency: infected mosquito to human

DurationPf
200; duration of infection (How many days does the infection last?)

LatentPf
10; latency (How many days after the infectious bite does the infection start?)

FeverPf
0.3; probability of fever

mnFeverPf
10; mean of Gaussian-distributed timing of fever incident (relative to start of the infection)

vrFeverPf
0.1; standard deviation of Gaussian-distributed timing of fever incident (relative to start of the infection)

TreatPf
0.5; probability of treatment

mnTreatPf
1; timing of treatment (relative to start of the fever)

mnChemoprophylaxisPf
32; Prophylaxis, time to susceptibility (duration of protection due to chemoprophylaxis)

PEProtectPf
0.99; proportion protected by PE vaccination (probability vaccination successful)

peBlockPf
1; proportion of infections blocked by PE vaccination

mnPEPf
270; mean duration of protection from PE vaccination

vrPEPf
50; standard deviation of protection from PE vaccination

GSProtectPf
1; proportion protected by GS vaccination (probability vaccination successful)

gsBlockPf
0.9; proportion of infections blocked by GS vaccination

mnGSPf
180; mean duration of protection from GS vaccination

vrGSPf
20; standard deviation of protection from GS vaccination

rdtSensPf
0.9; RDT sensitivity

rdtSpecPf
0.1; RDT specificity

lmSensPf
0.9; Light Microscopy sensitivity

lmSpecPf
0.1; Light Microscopy specificity

Value

Defines a field (list) PfSI_PAR in HumanPop and public methods in Human

Examples

PfSI.Setup()

#> initializing PfSI PATHOGEN module