Simulate Biting Process from Population Mean Number of Bites

Simulate a biting process to result in a negative-binomially distributed cumulative number of bites per person over days. The biting process for each person is a Poisson process with individual gamma-distributed rates. Because the negative binomial distribution arises as a compound Poisson-Gamma distribution, it will also arise from independent Poisson processes each with gamma-distributed rates, because the expected value of a Poisson process is also Poisson. This calls gammaRates.

SimBite_MeanBites(nH, meanNumberBites, days = 365, shape = 5, plot = TRUE)

Arguments

nH
number of humans

meanNumberBites
population mean number of bites (over time period equal to days)

days
number of days

shape
shape parameter of gamma-distributed variation in individual biting rate (higher values lead to more normally distributed cumulative biting counts)

plot
visualize biting distribution and negative binomial fit

Value

a list of length nH; each element is the arrivial times of bites on that individual