Generate Seasonal Emergence (Lambda) for Emerge model of Aquatic Ecology

Write me!

aquaEmerge_makeLambda(lambda, lambdaWeight = NULL, offset = NULL)

Arguments

lambda
vector of length equal to number of sites/patches where each element is the number of emerging adult females per human per day averaged over one year

lambdaWeight
vector of weights applied to each site

offset
seasonal offset in peak emergence

Value

does stuff

Examples

aquaEmerge_makeLambda(lambda = c(2,3,4))

#> [[1]]
#>   [1] 1.514267e+00 1.539884e+00 1.565485e+00 1.591064e+00 1.616613e+00
#>   [6] 1.642124e+00 1.667589e+00 1.693001e+00 1.718353e+00 1.743636e+00
#>  [11] 1.768844e+00 1.793969e+00 1.819004e+00 1.843941e+00 1.868772e+00
#>  [16] 1.893491e+00 1.918089e+00 1.942561e+00 1.966898e+00 1.991093e+00
#>  [21] 2.015139e+00 2.039030e+00 2.062757e+00 2.086314e+00 2.109694e+00
#>  [26] 2.132890e+00 2.155895e+00 2.178703e+00 2.201306e+00 2.223697e+00
#>  [31] 2.245871e+00 2.267821e+00 2.289539e+00 2.311021e+00 2.332258e+00
#>  [36] 2.353246e+00 2.373977e+00 2.394447e+00 2.414647e+00 2.434574e+00
#>  [41] 2.454220e+00 2.473579e+00 2.492647e+00 2.511418e+00 2.529885e+00
#>  [46] 2.548044e+00 2.565889e+00 2.583415e+00 2.600616e+00 2.617488e+00
#>  [51] 2.634026e+00 2.650224e+00 2.666077e+00 2.681582e+00 2.696734e+00
#>  [56] 2.711527e+00 2.725958e+00 2.740022e+00 2.753716e+00 2.767035e+00
#>  [61] 2.779974e+00 2.792532e+00 2.804702e+00 2.816483e+00 2.827871e+00
#>  [66] 2.838861e+00 2.849452e+00 2.859639e+00 2.869420e+00 2.878792e+00
#>  [71] 2.887752e+00 2.896297e+00 2.904425e+00 2.912134e+00 2.919421e+00
#>  [76] 2.926283e+00 2.932720e+00 2.938729e+00 2.944309e+00 2.949456e+00
#>  [81] 2.954172e+00 2.958452e+00 2.962298e+00 2.965706e+00 2.968677e+00
#>  [86] 2.971209e+00 2.973302e+00 2.974955e+00 2.976168e+00 2.976940e+00
#>  [91] 2.977271e+00 2.977160e+00 2.976609e+00 2.975617e+00 2.974184e+00
#>  [96] 2.972311e+00 2.969998e+00 2.967246e+00 2.964056e+00 2.960429e+00
#> [101] 2.956366e+00 2.951868e+00 2.946937e+00 2.941573e+00 2.935778e+00
#> [106] 2.929555e+00 2.922905e+00 2.915830e+00 2.908332e+00 2.900413e+00
#> [111] 2.892076e+00 2.883323e+00 2.874157e+00 2.864580e+00 2.854596e+00
#> [116] 2.844207e+00 2.833416e+00 2.822226e+00 2.810642e+00 2.798666e+00
#> [121] 2.786301e+00 2.773552e+00 2.760422e+00 2.746916e+00 2.733036e+00
#> [126] 2.718788e+00 2.704175e+00 2.689202e+00 2.673874e+00 2.658194e+00
#> [131] 2.642167e+00 2.625799e+00 2.609094e+00 2.592057e+00 2.574692e+00
#> [136] 2.557006e+00 2.539004e+00 2.520690e+00 2.502070e+00 2.483150e+00
#> [141] 2.463936e+00 2.444432e+00 2.424645e+00 2.404581e+00 2.384245e+00
#> [146] 2.363644e+00 2.342784e+00 2.321670e+00 2.300310e+00 2.278709e+00
#> [151] 2.256874e+00 2.234812e+00 2.212528e+00 2.190030e+00 2.167324e+00
#> [156] 2.144417e+00 2.121316e+00 2.098027e+00 2.074557e+00 2.050914e+00
#> [161] 2.027105e+00 2.003135e+00 1.979014e+00 1.954747e+00 1.930341e+00
#> [166] 1.905805e+00 1.881146e+00 1.856370e+00 1.831485e+00 1.806498e+00
#> [171] 1.781418e+00 1.756250e+00 1.731004e+00 1.705685e+00 1.680302e+00
#> [176] 1.654863e+00 1.629374e+00 1.603843e+00 1.578278e+00 1.552687e+00
#> [181] 1.527077e+00 1.501455e+00 1.475829e+00 1.450208e+00 1.424598e+00
#> [186] 1.399006e+00 1.373441e+00 1.347911e+00 1.322422e+00 1.296982e+00
#> [191] 1.271599e+00 1.246281e+00 1.221034e+00 1.195867e+00 1.170786e+00
#> [196] 1.145799e+00 1.120915e+00 1.096139e+00 1.071479e+00 1.046943e+00
#> [201] 1.022538e+00 9.982708e-01 9.741491e-01 9.501799e-01 9.263702e-01
#> [206] 9.027271e-01 8.792577e-01 8.559688e-01 8.328674e-01 8.099603e-01
#> [211] 7.872543e-01 7.647561e-01 7.424725e-01 7.204100e-01 6.985751e-01
#> [216] 6.769743e-01 6.556140e-01 6.345006e-01 6.136403e-01 5.930392e-01
#> [221] 5.727036e-01 5.526393e-01 5.328525e-01 5.133488e-01 4.941341e-01
#> [226] 4.752142e-01 4.565945e-01 4.382807e-01 4.202781e-01 4.025920e-01
#> [231] 3.852278e-01 3.681906e-01 3.514854e-01 3.351171e-01 3.190907e-01
#> [236] 3.034108e-01 2.880821e-01 2.731092e-01 2.584965e-01 2.442483e-01
#> [241] 2.303688e-01 2.168622e-01 2.037324e-01 1.909834e-01 1.786189e-01
#> [246] 1.666426e-01 1.550580e-01 1.438686e-01 1.330777e-01 1.226885e-01
#> [251] 1.127040e-01 1.031273e-01 9.396109e-02 8.520818e-02 7.687113e-02
#> [256] 6.895242e-02 6.145440e-02 5.437928e-02 4.772917e-02 4.150603e-02
#> [261] 3.571171e-02 3.034792e-02 2.541626e-02 2.091818e-02 1.685502e-02
#> [266] 1.322798e-02 1.003814e-02 7.286441e-03 4.973700e-03 3.100602e-03
#> [271] 1.667703e-03 6.754257e-04 1.240654e-04 1.378522e-05 3.446177e-04
#> [276] 1.116465e-03 2.329098e-03 3.982158e-03 6.075155e-03 8.607468e-03
#> [281] 1.157835e-02 1.498691e-02 1.883215e-02 2.311293e-02 2.782798e-02
#> [286] 3.297589e-02 3.855516e-02 4.456411e-02 5.100098e-02 5.786384e-02
#> [291] 6.515068e-02 7.285934e-02 8.098751e-02 8.953281e-02 9.849269e-02
#> [296] 1.078645e-01 1.176455e-01 1.278327e-01 1.384232e-01 1.494137e-01
#> [301] 1.608011e-01 1.725820e-01 1.847528e-01 1.973101e-01 2.102499e-01
#> [306] 2.235686e-01 2.372622e-01 2.513265e-01 2.657575e-01 2.805509e-01
#> [311] 2.957023e-01 3.112071e-01 3.270609e-01 3.432588e-01 3.597962e-01
#> [316] 3.766680e-01 3.938694e-01 4.113952e-01 4.292401e-01 4.473990e-01
#> [321] 4.658665e-01 4.846370e-01 5.037050e-01 5.230649e-01 5.427109e-01
#> [326] 5.626372e-01 5.828379e-01 6.033070e-01 6.240384e-01 6.450261e-01
#> [331] 6.662637e-01 6.877450e-01 7.094636e-01 7.314132e-01 7.535871e-01
#> [336] 7.759788e-01 7.985817e-01 8.213891e-01 8.443942e-01 8.675902e-01
#> [341] 8.909702e-01 9.145274e-01 9.382547e-01 9.621450e-01 9.861914e-01
#> [346] 1.010387e+00 1.034724e+00 1.059195e+00 1.083794e+00 1.108513e+00
#> [351] 1.133344e+00 1.158280e+00 1.183315e+00 1.208440e+00 1.233648e+00
#> [356] 1.258932e+00 1.284283e+00 1.309695e+00 1.335161e+00 1.360671e+00
#> [361] 1.386220e+00 1.411799e+00 1.437401e+00 1.463018e+00 1.488642e+00
#> 
#> [[2]]
#>   [1] 7.436560e-01 7.562364e-01 7.688095e-01 7.813713e-01 7.939182e-01
#>   [6] 8.064465e-01 8.189525e-01 8.314324e-01 8.438826e-01 8.562994e-01
#>  [11] 8.686790e-01 8.810179e-01 8.933123e-01 9.055587e-01 9.177534e-01
#>  [16] 9.298927e-01 9.419731e-01 9.539911e-01 9.659430e-01 9.778252e-01
#>  [21] 9.896344e-01 1.001367e+00 1.013019e+00 1.024588e+00 1.036070e+00
#>  [26] 1.047462e+00 1.058760e+00 1.069960e+00 1.081061e+00 1.092057e+00
#>  [31] 1.102947e+00 1.113726e+00 1.124392e+00 1.134942e+00 1.145371e+00
#>  [36] 1.155678e+00 1.165860e+00 1.175912e+00 1.185833e+00 1.195618e+00
#>  [41] 1.205267e+00 1.214774e+00 1.224138e+00 1.233357e+00 1.242426e+00
#>  [46] 1.251344e+00 1.260108e+00 1.268714e+00 1.277162e+00 1.285448e+00
#>  [51] 1.293569e+00 1.301524e+00 1.309310e+00 1.316924e+00 1.324365e+00
#>  [56] 1.331630e+00 1.338717e+00 1.345624e+00 1.352349e+00 1.358890e+00
#>  [61] 1.365245e+00 1.371411e+00 1.377389e+00 1.383174e+00 1.388767e+00
#>  [66] 1.394164e+00 1.399365e+00 1.404368e+00 1.409171e+00 1.413774e+00
#>  [71] 1.418174e+00 1.422371e+00 1.426362e+00 1.430148e+00 1.433727e+00
#>  [76] 1.437097e+00 1.440258e+00 1.443209e+00 1.445949e+00 1.448477e+00
#>  [81] 1.450793e+00 1.452895e+00 1.454783e+00 1.456457e+00 1.457916e+00
#>  [86] 1.459160e+00 1.460188e+00 1.461000e+00 1.461595e+00 1.461974e+00
#>  [91] 1.462137e+00 1.462083e+00 1.461812e+00 1.461325e+00 1.460621e+00
#>  [96] 1.459701e+00 1.458565e+00 1.457214e+00 1.455647e+00 1.453866e+00
#> [101] 1.451871e+00 1.449662e+00 1.447240e+00 1.444606e+00 1.441760e+00
#> [106] 1.438704e+00 1.435438e+00 1.431963e+00 1.428281e+00 1.424392e+00
#> [111] 1.420298e+00 1.415999e+00 1.411498e+00 1.406795e+00 1.401891e+00
#> [116] 1.396789e+00 1.391490e+00 1.385995e+00 1.380305e+00 1.374424e+00
#> [121] 1.368352e+00 1.362091e+00 1.355643e+00 1.349010e+00 1.342193e+00
#> [126] 1.335196e+00 1.328020e+00 1.320667e+00 1.313139e+00 1.305438e+00
#> [131] 1.297568e+00 1.289529e+00 1.281325e+00 1.272958e+00 1.264431e+00
#> [136] 1.255745e+00 1.246904e+00 1.237910e+00 1.228766e+00 1.219474e+00
#> [141] 1.210038e+00 1.200460e+00 1.190743e+00 1.180889e+00 1.170902e+00
#> [146] 1.160785e+00 1.150540e+00 1.140172e+00 1.129682e+00 1.119074e+00
#> [151] 1.108350e+00 1.097516e+00 1.086572e+00 1.075523e+00 1.064372e+00
#> [156] 1.053123e+00 1.041778e+00 1.030340e+00 1.018815e+00 1.007203e+00
#> [161] 9.955105e-01 9.837392e-01 9.718930e-01 9.599755e-01 9.479902e-01
#> [166] 9.359405e-01 9.238302e-01 9.116627e-01 8.994418e-01 8.871709e-01
#> [171] 8.748538e-01 8.624941e-01 8.500954e-01 8.376615e-01 8.251959e-01
#> [176] 8.127025e-01 8.001849e-01 7.876469e-01 7.750920e-01 7.625241e-01
#> [181] 7.499469e-01 7.373641e-01 7.247794e-01 7.121966e-01 6.996194e-01
#> [186] 6.870516e-01 6.744967e-01 6.619586e-01 6.494410e-01 6.369476e-01
#> [191] 6.244821e-01 6.120482e-01 5.996495e-01 5.872898e-01 5.749727e-01
#> [196] 5.627018e-01 5.504808e-01 5.383134e-01 5.262030e-01 5.141534e-01
#> [201] 5.021681e-01 4.902505e-01 4.784044e-01 4.666331e-01 4.549401e-01
#> [206] 4.433290e-01 4.318032e-01 4.203660e-01 4.090209e-01 3.977713e-01
#> [211] 3.866204e-01 3.755715e-01 3.646280e-01 3.537931e-01 3.430700e-01
#> [216] 3.324619e-01 3.219719e-01 3.116031e-01 3.013586e-01 2.912414e-01
#> [221] 2.812546e-01 2.714010e-01 2.616837e-01 2.521055e-01 2.426691e-01
#> [226] 2.333776e-01 2.242334e-01 2.152395e-01 2.063984e-01 1.977128e-01
#> [231] 1.891853e-01 1.808183e-01 1.726144e-01 1.645759e-01 1.567053e-01
#> [236] 1.490050e-01 1.414771e-01 1.341239e-01 1.269475e-01 1.199503e-01
#> [241] 1.131340e-01 1.065010e-01 1.000529e-01 9.379189e-02 8.771969e-02
#> [246] 8.183813e-02 7.614896e-02 7.065385e-02 6.535443e-02 6.025229e-02
#> [251] 5.534892e-02 5.064578e-02 4.614427e-02 4.184571e-02 3.775139e-02
#> [256] 3.386252e-02 3.018024e-02 2.670565e-02 2.343978e-02 2.038360e-02
#> [261] 1.753801e-02 1.490385e-02 1.248192e-02 1.027291e-02 8.277494e-03
#> [266] 6.496257e-03 4.929727e-03 3.578369e-03 2.442583e-03 1.522705e-03
#> [271] 8.190083e-04 3.317014e-04 6.092851e-05 6.769918e-06 1.692417e-04
#> [276] 5.482956e-04 1.143819e-03 1.955637e-03 2.983507e-03 4.227125e-03
#> [281] 5.686123e-03 7.360069e-03 9.248466e-03 1.135075e-02 1.366631e-02
#> [286] 1.619445e-02 1.893443e-02 2.188542e-02 2.504656e-02 2.841692e-02
#> [291] 3.199548e-02 3.578120e-02 3.977295e-02 4.396954e-02 4.836973e-02
#> [296] 5.297223e-02 5.777566e-02 6.277860e-02 6.797958e-02 7.337704e-02
#> [301] 7.896939e-02 8.475497e-02 9.073207e-02 9.689892e-02 1.032537e-01
#> [306] 1.097945e-01 1.165194e-01 1.234264e-01 1.305135e-01 1.377785e-01
#> [311] 1.452193e-01 1.528337e-01 1.606195e-01 1.685743e-01 1.766958e-01
#> [316] 1.849816e-01 1.934291e-01 2.020360e-01 2.107997e-01 2.197175e-01
#> [321] 2.287869e-01 2.380051e-01 2.473694e-01 2.568770e-01 2.665252e-01
#> [326] 2.763110e-01 2.862315e-01 2.962839e-01 3.064651e-01 3.167721e-01
#> [331] 3.272019e-01 3.377514e-01 3.484174e-01 3.591968e-01 3.700864e-01
#> [336] 3.810830e-01 3.921833e-01 4.033840e-01 4.146818e-01 4.260733e-01
#> [341] 4.375552e-01 4.491241e-01 4.607766e-01 4.725091e-01 4.843183e-01
#> [346] 4.962006e-01 5.081525e-01 5.201704e-01 5.322508e-01 5.443902e-01
#> [351] 5.565849e-01 5.688312e-01 5.811257e-01 5.934645e-01 6.058442e-01
#> [356] 6.182610e-01 6.307111e-01 6.431911e-01 6.556970e-01 6.682253e-01
#> [361] 6.807723e-01 6.933341e-01 7.059071e-01 7.184876e-01 7.310718e-01
#> 
#> [[3]]
#>   [1] 4.877862e-02 4.960381e-02 5.042851e-02 5.125248e-02 5.207547e-02
#>   [6] 5.289724e-02 5.371754e-02 5.453614e-02 5.535278e-02 5.616723e-02
#>  [11] 5.697925e-02 5.778860e-02 5.859502e-02 5.939830e-02 6.019819e-02
#>  [16] 6.099444e-02 6.178683e-02 6.257513e-02 6.335908e-02 6.413848e-02
#>  [21] 6.491308e-02 6.568265e-02 6.644697e-02 6.720581e-02 6.795894e-02
#>  [26] 6.870615e-02 6.944721e-02 7.018189e-02 7.090999e-02 7.163129e-02
#>  [31] 7.234557e-02 7.305263e-02 7.375224e-02 7.444421e-02 7.512833e-02
#>  [36] 7.580440e-02 7.647222e-02 7.713158e-02 7.778230e-02 7.842418e-02
#>  [41] 7.905703e-02 7.968067e-02 8.029490e-02 8.089955e-02 8.149443e-02
#>  [46] 8.207938e-02 8.265422e-02 8.321877e-02 8.377287e-02 8.431636e-02
#>  [51] 8.484907e-02 8.537085e-02 8.588154e-02 8.638100e-02 8.686906e-02
#>  [56] 8.734560e-02 8.781046e-02 8.826351e-02 8.870462e-02 8.913365e-02
#>  [61] 8.955047e-02 8.995498e-02 9.034703e-02 9.072653e-02 9.109334e-02
#>  [66] 9.144738e-02 9.178853e-02 9.211669e-02 9.243176e-02 9.273365e-02
#>  [71] 9.302227e-02 9.329754e-02 9.355937e-02 9.380768e-02 9.404241e-02
#>  [76] 9.426349e-02 9.447083e-02 9.466440e-02 9.484412e-02 9.500995e-02
#>  [81] 9.516183e-02 9.529973e-02 9.542360e-02 9.553339e-02 9.562909e-02
#>  [86] 9.571067e-02 9.577809e-02 9.583134e-02 9.587040e-02 9.589526e-02
#>  [91] 9.590592e-02 9.590237e-02 9.588461e-02 9.585264e-02 9.580649e-02
#>  [96] 9.574615e-02 9.567165e-02 9.558301e-02 9.548026e-02 9.536342e-02
#> [101] 9.523253e-02 9.508764e-02 9.492878e-02 9.475599e-02 9.456934e-02
#> [106] 9.436888e-02 9.415466e-02 9.392675e-02 9.368522e-02 9.343014e-02
#> [111] 9.316158e-02 9.287962e-02 9.258436e-02 9.227586e-02 9.195424e-02
#> [116] 9.161957e-02 9.127197e-02 9.091153e-02 9.053836e-02 9.015257e-02
#> [121] 8.975427e-02 8.934359e-02 8.892065e-02 8.848556e-02 8.803847e-02
#> [126] 8.757950e-02 8.710878e-02 8.662646e-02 8.613268e-02 8.562759e-02
#> [131] 8.511134e-02 8.458407e-02 8.404595e-02 8.349714e-02 8.293779e-02
#> [136] 8.236807e-02 8.178816e-02 8.119822e-02 8.059843e-02 7.998897e-02
#> [141] 7.937001e-02 7.874175e-02 7.810436e-02 7.745804e-02 7.680297e-02
#> [146] 7.613936e-02 7.546739e-02 7.478727e-02 7.409920e-02 7.340338e-02
#> [151] 7.270002e-02 7.198932e-02 7.127151e-02 7.054678e-02 6.981536e-02
#> [156] 6.907746e-02 6.833330e-02 6.758310e-02 6.682709e-02 6.606548e-02
#> [161] 6.529851e-02 6.452639e-02 6.374937e-02 6.296766e-02 6.218151e-02
#> [166] 6.139113e-02 6.059678e-02 5.979868e-02 5.899707e-02 5.819219e-02
#> [171] 5.738427e-02 5.657356e-02 5.576030e-02 5.494472e-02 5.412707e-02
#> [176] 5.330759e-02 5.248652e-02 5.166411e-02 5.084060e-02 5.001624e-02
#> [181] 4.919126e-02 4.836591e-02 4.754045e-02 4.671511e-02 4.589013e-02
#> [186] 4.506576e-02 4.424225e-02 4.341984e-02 4.259878e-02 4.177930e-02
#> [191] 4.096165e-02 4.014607e-02 3.933280e-02 3.852209e-02 3.771417e-02
#> [196] 3.690929e-02 3.610768e-02 3.530958e-02 3.451523e-02 3.372486e-02
#> [201] 3.293870e-02 3.215700e-02 3.137997e-02 3.060786e-02 2.984088e-02
#> [206] 2.907928e-02 2.832326e-02 2.757306e-02 2.682890e-02 2.609101e-02
#> [211] 2.535958e-02 2.463486e-02 2.391704e-02 2.320635e-02 2.250299e-02
#> [216] 2.180717e-02 2.111910e-02 2.043898e-02 1.976701e-02 1.910339e-02
#> [221] 1.844833e-02 1.780200e-02 1.716462e-02 1.653635e-02 1.591739e-02
#> [226] 1.530793e-02 1.470814e-02 1.411820e-02 1.353829e-02 1.296858e-02
#> [231] 1.240923e-02 1.186041e-02 1.132229e-02 1.079503e-02 1.027877e-02
#> [236] 9.773680e-03 9.279902e-03 8.797584e-03 8.326869e-03 7.867895e-03
#> [241] 7.420800e-03 6.985716e-03 6.562770e-03 6.152090e-03 5.753797e-03
#> [246] 5.368008e-03 4.994837e-03 4.634397e-03 4.286792e-03 3.952127e-03
#> [251] 3.630501e-03 3.322008e-03 3.026740e-03 2.744785e-03 2.476227e-03
#> [256] 2.221144e-03 1.979612e-03 1.751703e-03 1.537485e-03 1.337021e-03
#> [261] 1.150370e-03 9.775884e-04 8.187262e-04 6.738309e-04 5.429456e-04
#> [266] 4.261089e-04 3.233555e-04 2.347159e-04 1.602163e-04 9.987876e-05
#> [271] 5.372120e-05 2.175728e-05 3.996483e-06 4.440591e-07 1.110107e-05
#> [276] 3.596435e-05 7.502653e-05 1.282761e-04 1.956971e-04 2.772698e-04
#> [281] 3.729698e-04 4.827689e-04 6.066345e-04 7.445299e-04 8.964142e-04
#> [286] 1.062243e-03 1.241966e-03 1.435530e-03 1.642879e-03 1.863951e-03
#> [291] 2.098679e-03 2.346996e-03 2.608826e-03 2.884093e-03 3.172715e-03
#> [296] 3.474607e-03 3.789678e-03 4.117836e-03 4.458984e-03 4.813019e-03
#> [301] 5.179838e-03 5.559332e-03 5.951388e-03 6.355890e-03 6.772718e-03
#> [306] 7.201748e-03 7.642855e-03 8.095906e-03 8.560767e-03 9.037302e-03
#> [311] 9.525368e-03 1.002482e-02 1.053551e-02 1.105729e-02 1.159001e-02
#> [316] 1.213349e-02 1.268760e-02 1.325215e-02 1.382698e-02 1.441193e-02
#> [321] 1.500682e-02 1.561146e-02 1.622570e-02 1.684933e-02 1.748218e-02
#> [326] 1.812406e-02 1.877478e-02 1.943414e-02 2.010196e-02 2.077803e-02
#> [331] 2.146215e-02 2.215412e-02 2.285374e-02 2.356079e-02 2.427507e-02
#> [336] 2.499637e-02 2.572447e-02 2.645916e-02 2.720022e-02 2.794742e-02
#> [341] 2.870056e-02 2.945939e-02 3.022371e-02 3.099329e-02 3.176789e-02
#> [346] 3.254728e-02 3.333124e-02 3.411953e-02 3.491192e-02 3.570818e-02
#> [351] 3.650806e-02 3.731134e-02 3.811777e-02 3.892711e-02 3.973913e-02
#> [356] 4.055358e-02 4.137023e-02 4.218882e-02 4.300913e-02 4.383090e-02
#> [361] 4.465389e-02 4.547785e-02 4.630256e-02 4.712775e-02 4.795318e-02
#>