MACRO: Generate Seasonal Emergence (Lambda) for Emerge model of Aquatic Ecology

Generate lambda for all patches.

makeLambda_Macro(aquaPars)

Arguments

aquaPars
a list of the following structure

lambda: vector of length equal to number of patches MacroPatch where each element is the number of emerging adult females per human per day averaged over one year (required)

lambdaWeight: vector of weights applied to each site (if not specified or set to NULL initialize to Gamma(1,1) distribution)

offset: vector of seasonal offsets in peak emergence applied to each site (if not specified or set to NULL initialize to 0 for all sites)

Value

list lambda where each element is the daily emergence for that MacroPatch

Examples

makeLambda_Macro(aquaPars = list(lambda=c(5,10,15)))

#> [[1]]
#>   [1] 2.791980e-01 2.839212e-01 2.886416e-01 2.933578e-01 2.980684e-01
#>   [6] 3.027720e-01 3.074673e-01 3.121527e-01 3.168270e-01 3.214888e-01
#>  [11] 3.261366e-01 3.307691e-01 3.353849e-01 3.399827e-01 3.445610e-01
#>  [16] 3.491186e-01 3.536541e-01 3.581661e-01 3.626533e-01 3.671144e-01
#>  [21] 3.715480e-01 3.759529e-01 3.803277e-01 3.846711e-01 3.889819e-01
#>  [26] 3.932587e-01 3.975004e-01 4.017055e-01 4.058730e-01 4.100016e-01
#>  [31] 4.140900e-01 4.181370e-01 4.221414e-01 4.261021e-01 4.300179e-01
#>  [36] 4.338875e-01 4.377100e-01 4.414840e-01 4.452086e-01 4.488826e-01
#>  [41] 4.525049e-01 4.560744e-01 4.595901e-01 4.630510e-01 4.664560e-01
#>  [46] 4.698041e-01 4.730943e-01 4.763257e-01 4.794973e-01 4.826081e-01
#>  [51] 4.856572e-01 4.886438e-01 4.915669e-01 4.944256e-01 4.972192e-01
#>  [56] 4.999468e-01 5.026075e-01 5.052007e-01 5.077255e-01 5.101812e-01
#>  [61] 5.125670e-01 5.148823e-01 5.171263e-01 5.192985e-01 5.213980e-01
#>  [66] 5.234245e-01 5.253771e-01 5.272554e-01 5.290588e-01 5.307868e-01
#>  [71] 5.324388e-01 5.340144e-01 5.355130e-01 5.369343e-01 5.382779e-01
#>  [76] 5.395432e-01 5.407300e-01 5.418380e-01 5.428667e-01 5.438158e-01
#>  [81] 5.446852e-01 5.454745e-01 5.461834e-01 5.468119e-01 5.473597e-01
#>  [86] 5.478266e-01 5.482125e-01 5.485173e-01 5.487408e-01 5.488832e-01
#>  [91] 5.489442e-01 5.489238e-01 5.488222e-01 5.486392e-01 5.483750e-01
#>  [96] 5.480297e-01 5.476032e-01 5.470959e-01 5.465077e-01 5.458390e-01
#> [101] 5.450898e-01 5.442605e-01 5.433512e-01 5.423622e-01 5.412939e-01
#> [106] 5.401465e-01 5.389203e-01 5.376158e-01 5.362334e-01 5.347733e-01
#> [111] 5.332362e-01 5.316223e-01 5.299323e-01 5.281665e-01 5.263256e-01
#> [116] 5.244100e-01 5.224204e-01 5.203574e-01 5.182214e-01 5.160132e-01
#> [121] 5.137335e-01 5.113829e-01 5.089620e-01 5.064717e-01 5.039126e-01
#> [126] 5.012856e-01 4.985913e-01 4.958306e-01 4.930043e-01 4.901133e-01
#> [131] 4.871584e-01 4.841404e-01 4.810603e-01 4.779190e-01 4.747175e-01
#> [136] 4.714565e-01 4.681372e-01 4.647606e-01 4.613275e-01 4.578391e-01
#> [141] 4.542963e-01 4.507003e-01 4.470520e-01 4.433526e-01 4.396031e-01
#> [146] 4.358047e-01 4.319585e-01 4.280657e-01 4.241273e-01 4.201446e-01
#> [151] 4.161187e-01 4.120509e-01 4.079422e-01 4.037941e-01 3.996076e-01
#> [156] 3.953840e-01 3.911246e-01 3.868306e-01 3.825034e-01 3.781441e-01
#> [161] 3.737541e-01 3.693347e-01 3.648872e-01 3.604129e-01 3.559131e-01
#> [166] 3.513892e-01 3.468425e-01 3.422744e-01 3.376861e-01 3.330791e-01
#> [171] 3.284548e-01 3.238145e-01 3.191595e-01 3.144913e-01 3.098113e-01
#> [176] 3.051208e-01 3.004212e-01 2.957139e-01 2.910003e-01 2.862818e-01
#> [181] 2.815598e-01 2.768357e-01 2.721110e-01 2.673869e-01 2.626649e-01
#> [186] 2.579464e-01 2.532328e-01 2.485255e-01 2.438259e-01 2.391354e-01
#> [191] 2.344554e-01 2.297872e-01 2.251322e-01 2.204919e-01 2.158675e-01
#> [196] 2.112606e-01 2.066723e-01 2.021042e-01 1.975575e-01 1.930336e-01
#> [201] 1.885338e-01 1.840595e-01 1.796120e-01 1.751926e-01 1.708026e-01
#> [206] 1.664433e-01 1.621160e-01 1.578221e-01 1.535627e-01 1.493391e-01
#> [211] 1.451526e-01 1.410045e-01 1.368958e-01 1.328280e-01 1.288021e-01
#> [216] 1.248194e-01 1.208810e-01 1.169882e-01 1.131420e-01 1.093436e-01
#> [221] 1.055941e-01 1.018947e-01 9.824644e-02 9.465039e-02 9.110762e-02
#> [226] 8.761919e-02 8.418613e-02 8.080945e-02 7.749016e-02 7.422924e-02
#> [231] 7.102766e-02 6.788636e-02 6.480628e-02 6.178833e-02 5.883340e-02
#> [236] 5.594237e-02 5.311610e-02 5.035542e-02 4.766114e-02 4.503408e-02
#> [241] 4.247501e-02 3.998468e-02 3.756384e-02 3.521320e-02 3.293345e-02
#> [246] 3.072528e-02 2.858934e-02 2.652626e-02 2.453665e-02 2.262110e-02
#> [251] 2.078018e-02 1.901443e-02 1.732439e-02 1.571054e-02 1.417337e-02
#> [256] 1.271333e-02 1.133086e-02 1.002636e-02 8.800224e-03 7.652812e-03
#> [261] 6.584464e-03 5.595498e-03 4.686206e-03 3.856858e-03 3.107700e-03
#> [266] 2.438953e-03 1.850815e-03 1.343462e-03 9.170424e-04 5.716838e-04
#> [271] 3.074882e-04 1.245339e-04 2.287498e-05 2.541696e-06 6.354004e-05
#> [276] 2.058519e-04 4.294352e-04 7.342236e-04 1.120127e-03 1.587030e-03
#> [281] 2.134796e-03 2.763262e-03 3.472241e-03 4.261523e-03 5.130876e-03
#> [286] 6.080040e-03 7.108735e-03 8.216656e-03 9.403474e-03 1.066884e-02
#> [291] 1.201237e-02 1.343368e-02 1.493234e-02 1.650791e-02 1.815992e-02
#> [296] 1.988788e-02 2.169128e-02 2.356958e-02 2.552223e-02 2.754865e-02
#> [301] 2.964824e-02 3.182038e-02 3.406442e-02 3.637970e-02 3.876553e-02
#> [306] 4.122121e-02 4.374600e-02 4.633916e-02 4.899993e-02 5.172751e-02
#> [311] 5.452109e-02 5.737985e-02 6.030293e-02 6.328948e-02 6.633861e-02
#> [316] 6.944942e-02 7.262098e-02 7.585235e-02 7.914257e-02 8.249068e-02
#> [321] 8.589567e-02 8.935655e-02 9.287228e-02 9.644182e-02 1.000641e-01
#> [326] 1.037381e-01 1.074627e-01 1.112367e-01 1.150592e-01 1.189288e-01
#> [331] 1.228446e-01 1.268053e-01 1.308097e-01 1.348567e-01 1.389451e-01
#> [336] 1.430737e-01 1.472412e-01 1.514463e-01 1.556880e-01 1.599648e-01
#> [341] 1.642756e-01 1.686190e-01 1.729938e-01 1.773987e-01 1.818323e-01
#> [346] 1.862934e-01 1.907806e-01 1.952926e-01 1.998281e-01 2.043857e-01
#> [351] 2.089640e-01 2.135618e-01 2.181776e-01 2.228101e-01 2.274579e-01
#> [356] 2.321197e-01 2.367940e-01 2.414794e-01 2.461747e-01 2.508783e-01
#> [361] 2.555889e-01 2.603051e-01 2.650255e-01 2.697487e-01 2.744733e-01
#> 
#> [[2]]
#>   [1] 9.568813e+00 9.730690e+00 9.892470e+00 1.005411e+01 1.021555e+01
#>   [6] 1.037676e+01 1.053767e+01 1.069826e+01 1.085846e+01 1.101822e+01
#>  [11] 1.117752e+01 1.133628e+01 1.149448e+01 1.165206e+01 1.180897e+01
#>  [16] 1.196517e+01 1.212061e+01 1.227525e+01 1.242904e+01 1.258193e+01
#>  [21] 1.273388e+01 1.288485e+01 1.303478e+01 1.318364e+01 1.333138e+01
#>  [26] 1.347796e+01 1.362333e+01 1.376746e+01 1.391029e+01 1.405178e+01
#>  [31] 1.419190e+01 1.433060e+01 1.446784e+01 1.460359e+01 1.473779e+01
#>  [36] 1.487041e+01 1.500142e+01 1.513076e+01 1.525841e+01 1.538433e+01
#>  [41] 1.550848e+01 1.563081e+01 1.575131e+01 1.586992e+01 1.598662e+01
#>  [46] 1.610136e+01 1.621413e+01 1.632488e+01 1.643357e+01 1.654019e+01
#>  [51] 1.664469e+01 1.674705e+01 1.684723e+01 1.694520e+01 1.704095e+01
#>  [56] 1.713443e+01 1.722562e+01 1.731449e+01 1.740102e+01 1.748519e+01
#>  [61] 1.756696e+01 1.764631e+01 1.772321e+01 1.779766e+01 1.786962e+01
#>  [66] 1.793907e+01 1.800599e+01 1.807036e+01 1.813217e+01 1.819139e+01
#>  [71] 1.824801e+01 1.830201e+01 1.835337e+01 1.840208e+01 1.844813e+01
#>  [76] 1.849150e+01 1.853217e+01 1.857014e+01 1.860540e+01 1.863793e+01
#>  [81] 1.866773e+01 1.869478e+01 1.871907e+01 1.874061e+01 1.875939e+01
#>  [86] 1.877539e+01 1.878861e+01 1.879906e+01 1.880672e+01 1.881160e+01
#>  [91] 1.881369e+01 1.881299e+01 1.880951e+01 1.880324e+01 1.879419e+01
#>  [96] 1.878235e+01 1.876773e+01 1.875035e+01 1.873019e+01 1.870727e+01
#> [101] 1.868159e+01 1.865317e+01 1.862201e+01 1.858811e+01 1.855150e+01
#> [106] 1.851217e+01 1.847015e+01 1.842544e+01 1.837806e+01 1.832802e+01
#> [111] 1.827534e+01 1.822003e+01 1.816211e+01 1.810159e+01 1.803850e+01
#> [116] 1.797285e+01 1.790466e+01 1.783395e+01 1.776075e+01 1.768507e+01
#> [121] 1.760693e+01 1.752637e+01 1.744340e+01 1.735805e+01 1.727035e+01
#> [126] 1.718031e+01 1.708797e+01 1.699336e+01 1.689649e+01 1.679741e+01
#> [131] 1.669614e+01 1.659270e+01 1.648714e+01 1.637948e+01 1.626976e+01
#> [136] 1.615800e+01 1.604424e+01 1.592851e+01 1.581085e+01 1.569129e+01
#> [141] 1.556987e+01 1.544663e+01 1.532159e+01 1.519480e+01 1.506630e+01
#> [146] 1.493612e+01 1.480430e+01 1.467088e+01 1.453591e+01 1.439941e+01
#> [151] 1.426143e+01 1.412202e+01 1.398120e+01 1.383903e+01 1.369555e+01
#> [156] 1.355080e+01 1.340482e+01 1.325766e+01 1.310935e+01 1.295995e+01
#> [161] 1.280949e+01 1.265803e+01 1.250560e+01 1.235225e+01 1.219803e+01
#> [166] 1.204299e+01 1.188716e+01 1.173060e+01 1.157335e+01 1.141546e+01
#> [171] 1.125697e+01 1.109793e+01 1.093840e+01 1.077841e+01 1.061801e+01
#> [176] 1.045725e+01 1.029619e+01 1.013486e+01 9.973309e+00 9.811595e+00
#> [181] 9.649761e+00 9.487854e+00 9.325924e+00 9.164018e+00 9.002184e+00
#> [186] 8.840470e+00 8.678923e+00 8.517593e+00 8.356525e+00 8.195769e+00
#> [191] 8.035372e+00 7.875382e+00 7.715845e+00 7.556809e+00 7.398322e+00
#> [196] 7.240429e+00 7.083179e+00 6.926617e+00 6.770790e+00 6.615745e+00
#> [201] 6.461526e+00 6.308180e+00 6.155753e+00 6.004288e+00 5.853832e+00
#> [206] 5.704429e+00 5.556123e+00 5.408958e+00 5.262978e+00 5.118226e+00
#> [211] 4.974744e+00 4.832576e+00 4.691763e+00 4.552348e+00 4.414371e+00
#> [216] 4.277873e+00 4.142895e+00 4.009477e+00 3.877659e+00 3.747478e+00
#> [221] 3.618975e+00 3.492187e+00 3.367152e+00 3.243906e+00 3.122486e+00
#> [226] 3.002929e+00 2.885270e+00 2.769542e+00 2.655782e+00 2.544022e+00
#> [231] 2.434296e+00 2.326636e+00 2.221074e+00 2.117641e+00 2.016368e+00
#> [236] 1.917285e+00 1.820422e+00 1.725806e+00 1.633467e+00 1.543431e+00
#> [241] 1.455725e+00 1.370375e+00 1.287407e+00 1.206844e+00 1.128712e+00
#> [246] 1.053032e+00 9.798283e-01 9.091213e-01 8.409324e-01 7.752817e-01
#> [251] 7.121888e-01 6.516723e-01 5.937502e-01 5.384396e-01 4.857569e-01
#> [256] 4.357177e-01 3.883369e-01 3.436285e-01 3.016057e-01 2.622810e-01
#> [261] 2.256661e-01 1.917717e-01 1.606080e-01 1.321842e-01 1.065087e-01
#> [266] 8.358902e-02 6.343207e-02 4.604380e-02 3.142934e-02 1.959304e-02
#> [271] 1.053839e-02 4.268087e-03 7.839828e-04 8.711028e-05 2.177676e-03
#> [276] 7.055061e-03 1.471782e-02 2.516368e-02 3.838955e-02 5.439151e-02
#> [281] 7.316482e-02 9.470390e-02 1.190024e-01 1.460531e-01 1.758480e-01
#> [286] 2.083782e-01 2.436342e-01 2.816054e-01 3.222806e-01 3.656478e-01
#> [291] 4.116941e-01 4.604059e-01 5.117688e-01 5.657674e-01 6.223859e-01
#> [296] 6.816074e-01 7.434143e-01 8.077885e-01 8.747108e-01 9.441613e-01
#> [301] 1.016120e+00 1.090564e+00 1.167473e+00 1.246823e+00 1.328592e+00
#> [306] 1.412754e+00 1.499285e+00 1.588159e+00 1.679350e+00 1.772831e+00
#> [311] 1.868574e+00 1.966551e+00 2.066733e+00 2.169089e+00 2.273591e+00
#> [316] 2.380206e+00 2.488903e+00 2.599650e+00 2.712414e+00 2.827162e+00
#> [321] 2.943860e+00 3.062473e+00 3.182966e+00 3.305303e+00 3.429448e+00
#> [326] 3.555364e+00 3.683015e+00 3.812361e+00 3.943365e+00 4.075989e+00
#> [331] 4.210191e+00 4.345934e+00 4.483177e+00 4.621878e+00 4.761997e+00
#> [336] 4.903493e+00 5.046323e+00 5.190446e+00 5.335817e+00 5.482395e+00
#> [341] 5.630136e+00 5.778996e+00 5.928931e+00 6.079897e+00 6.231849e+00
#> [346] 6.384741e+00 6.538529e+00 6.693167e+00 6.848609e+00 7.004809e+00
#> [351] 7.161721e+00 7.319298e+00 7.477494e+00 7.636261e+00 7.795554e+00
#> [356] 7.955323e+00 8.115523e+00 8.276106e+00 8.437023e+00 8.598228e+00
#> [361] 8.759673e+00 8.921309e+00 9.083089e+00 9.244965e+00 9.406889e+00
#> 
#> [[3]]
#>   [1] 6.738628e-02 6.852626e-02 6.966556e-02 7.080385e-02 7.194079e-02
#>   [6] 7.307604e-02 7.420927e-02 7.534013e-02 7.646831e-02 7.759345e-02
#>  [11] 7.871523e-02 7.983331e-02 8.094737e-02 8.205708e-02 8.316209e-02
#>  [16] 8.426210e-02 8.535677e-02 8.644577e-02 8.752879e-02 8.860550e-02
#>  [21] 8.967559e-02 9.073873e-02 9.179461e-02 9.284293e-02 9.388336e-02
#>  [26] 9.491561e-02 9.593935e-02 9.695430e-02 9.796015e-02 9.895661e-02
#>  [31] 9.994337e-02 1.009201e-01 1.018866e-01 1.028426e-01 1.037877e-01
#>  [36] 1.047216e-01 1.056442e-01 1.065551e-01 1.074541e-01 1.083408e-01
#>  [41] 1.092151e-01 1.100766e-01 1.109251e-01 1.117604e-01 1.125823e-01
#>  [46] 1.133904e-01 1.141845e-01 1.149644e-01 1.157299e-01 1.164807e-01
#>  [51] 1.172166e-01 1.179374e-01 1.186429e-01 1.193329e-01 1.200072e-01
#>  [56] 1.206655e-01 1.213077e-01 1.219335e-01 1.225429e-01 1.231356e-01
#>  [61] 1.237115e-01 1.242703e-01 1.248119e-01 1.253361e-01 1.258429e-01
#>  [66] 1.263320e-01 1.268033e-01 1.272566e-01 1.276919e-01 1.281089e-01
#>  [71] 1.285076e-01 1.288879e-01 1.292496e-01 1.295927e-01 1.299169e-01
#>  [76] 1.302223e-01 1.305088e-01 1.307762e-01 1.310245e-01 1.312536e-01
#>  [81] 1.314634e-01 1.316539e-01 1.318250e-01 1.319767e-01 1.321089e-01
#>  [86] 1.322216e-01 1.323147e-01 1.323883e-01 1.324423e-01 1.324766e-01
#>  [91] 1.324913e-01 1.324864e-01 1.324619e-01 1.324177e-01 1.323540e-01
#>  [96] 1.322706e-01 1.321677e-01 1.320452e-01 1.319033e-01 1.317419e-01
#> [101] 1.315611e-01 1.313609e-01 1.311414e-01 1.309027e-01 1.306449e-01
#> [106] 1.303679e-01 1.300720e-01 1.297572e-01 1.294235e-01 1.290711e-01
#> [111] 1.287001e-01 1.283106e-01 1.279027e-01 1.274765e-01 1.270322e-01
#> [116] 1.265699e-01 1.260896e-01 1.255917e-01 1.250762e-01 1.245432e-01
#> [121] 1.239930e-01 1.234257e-01 1.228414e-01 1.222403e-01 1.216227e-01
#> [126] 1.209886e-01 1.203383e-01 1.196720e-01 1.189899e-01 1.182921e-01
#> [131] 1.175789e-01 1.168505e-01 1.161071e-01 1.153489e-01 1.145762e-01
#> [136] 1.137892e-01 1.129880e-01 1.121731e-01 1.113445e-01 1.105025e-01
#> [141] 1.096474e-01 1.087795e-01 1.078990e-01 1.070061e-01 1.061011e-01
#> [146] 1.051844e-01 1.042561e-01 1.033165e-01 1.023659e-01 1.014047e-01
#> [151] 1.004330e-01 9.945122e-02 9.845957e-02 9.745839e-02 9.644795e-02
#> [156] 9.542856e-02 9.440053e-02 9.336415e-02 9.231974e-02 9.126760e-02
#> [161] 9.020804e-02 8.914139e-02 8.806795e-02 8.698805e-02 8.590200e-02
#> [166] 8.481012e-02 8.371274e-02 8.261019e-02 8.150279e-02 8.039087e-02
#> [171] 7.927475e-02 7.815478e-02 7.703128e-02 7.590458e-02 7.477502e-02
#> [176] 7.364293e-02 7.250865e-02 7.137251e-02 7.023485e-02 6.909602e-02
#> [181] 6.795634e-02 6.681615e-02 6.567579e-02 6.453560e-02 6.339592e-02
#> [186] 6.225708e-02 6.111943e-02 5.998329e-02 5.884901e-02 5.771692e-02
#> [191] 5.658736e-02 5.546066e-02 5.433716e-02 5.321718e-02 5.210107e-02
#> [196] 5.098915e-02 4.988174e-02 4.877919e-02 4.768182e-02 4.658994e-02
#> [201] 4.550389e-02 4.442398e-02 4.335055e-02 4.228389e-02 4.122434e-02
#> [206] 4.017220e-02 3.912779e-02 3.809141e-02 3.706337e-02 3.604399e-02
#> [211] 3.503355e-02 3.403236e-02 3.304072e-02 3.205892e-02 3.108724e-02
#> [216] 3.012599e-02 2.917544e-02 2.823587e-02 2.730757e-02 2.639080e-02
#> [221] 2.548584e-02 2.459297e-02 2.371243e-02 2.284450e-02 2.198943e-02
#> [226] 2.114747e-02 2.031888e-02 1.950390e-02 1.870277e-02 1.791572e-02
#> [231] 1.714300e-02 1.638483e-02 1.564143e-02 1.491302e-02 1.419983e-02
#> [236] 1.350206e-02 1.281992e-02 1.215361e-02 1.150333e-02 1.086928e-02
#> [241] 1.025163e-02 9.650569e-03 9.066282e-03 8.498939e-03 7.948708e-03
#> [246] 7.415751e-03 6.900227e-03 6.402288e-03 5.922083e-03 5.459752e-03
#> [251] 5.015434e-03 4.589260e-03 4.181356e-03 3.791843e-03 3.420837e-03
#> [256] 3.068447e-03 2.734778e-03 2.419929e-03 2.123993e-03 1.847057e-03
#> [261] 1.589204e-03 1.350511e-03 1.131047e-03 9.308784e-04 7.500640e-04
#> [266] 5.886574e-04 4.467065e-04 3.242534e-04 2.213343e-04 1.379797e-04
#> [271] 7.421433e-05 3.005708e-05 5.521028e-06 6.134551e-07 1.533581e-05
#> [276] 4.968373e-05 1.036470e-04 1.772097e-04 2.703501e-04 3.830404e-04
#> [281] 5.152473e-04 6.669316e-04 8.380484e-04 1.028547e-03 1.238371e-03
#> [286] 1.467458e-03 1.715740e-03 1.983144e-03 2.269591e-03 2.574995e-03
#> [291] 2.899266e-03 3.242308e-03 3.604020e-03 3.984294e-03 4.383017e-03
#> [296] 4.800071e-03 5.235334e-03 5.688675e-03 6.159960e-03 6.649050e-03
#> [301] 7.155800e-03 7.680060e-03 8.221674e-03 8.780482e-03 9.356318e-03
#> [306] 9.949012e-03 1.055839e-02 1.118427e-02 1.182646e-02 1.248478e-02
#> [311] 1.315903e-02 1.384901e-02 1.455451e-02 1.527534e-02 1.601127e-02
#> [316] 1.676208e-02 1.752756e-02 1.830747e-02 1.910159e-02 1.990967e-02
#> [321] 2.073149e-02 2.156680e-02 2.241534e-02 2.327687e-02 2.415114e-02
#> [326] 2.503788e-02 2.593683e-02 2.684772e-02 2.777029e-02 2.870426e-02
#> [331] 2.964936e-02 3.060529e-02 3.157179e-02 3.254857e-02 3.353533e-02
#> [336] 3.453178e-02 3.553763e-02 3.655258e-02 3.757633e-02 3.860857e-02
#> [341] 3.964901e-02 4.069732e-02 4.175321e-02 4.281635e-02 4.388644e-02
#> [346] 4.496315e-02 4.604617e-02 4.713517e-02 4.822984e-02 4.932984e-02
#> [351] 5.043486e-02 5.154456e-02 5.265862e-02 5.377671e-02 5.489849e-02
#> [356] 5.602363e-02 5.715180e-02 5.828267e-02 5.941590e-02 6.055115e-02
#> [361] 6.168808e-02 6.282637e-02 6.396568e-02 6.510565e-02 6.624597e-02
#>