MACRO: Generate Seasonal Emergence (Lambda) for Emerge model of Aquatic Ecology

Generate lambda for all patches.

makeLambda_Macro(aquaPars)

Arguments

aquaPars
a list of the following structure

lambda: vector of length equal to number of patches MacroPatch where each element is the number of emerging adult females per human per day averaged over one year (required)

lambdaWeight: vector of weights applied to each site (if not specified or set to NULL initialize to Gamma(1,1) distribution)

offset: vector of seasonal offsets in peak emergence applied to each site (if not specified or set to NULL initialize to 0 for all sites)

Value

list lambda where each element is the daily emergence for that MacroPatch

Examples

makeLambda_Macro(aquaPars = list(lambda=c(5,10,15)))

#> [[1]]
#>   [1] 8.979753e-01 9.131664e-01 9.283485e-01 9.435171e-01 9.586677e-01
#>   [6] 9.737958e-01 9.888970e-01 1.003967e+00 1.019000e+00 1.033994e+00
#>  [11] 1.048942e+00 1.063842e+00 1.078688e+00 1.093475e+00 1.108200e+00
#>  [16] 1.122859e+00 1.137446e+00 1.151958e+00 1.166390e+00 1.180738e+00
#>  [21] 1.194998e+00 1.209165e+00 1.223236e+00 1.237205e+00 1.251070e+00
#>  [26] 1.264825e+00 1.278468e+00 1.291993e+00 1.305396e+00 1.318675e+00
#>  [31] 1.331824e+00 1.344840e+00 1.357720e+00 1.370458e+00 1.383053e+00
#>  [36] 1.395498e+00 1.407792e+00 1.419931e+00 1.431910e+00 1.443726e+00
#>  [41] 1.455377e+00 1.466857e+00 1.478165e+00 1.489296e+00 1.500247e+00
#>  [46] 1.511016e+00 1.521598e+00 1.531991e+00 1.542192e+00 1.552197e+00
#>  [51] 1.562004e+00 1.571609e+00 1.581011e+00 1.590205e+00 1.599190e+00
#>  [56] 1.607963e+00 1.616520e+00 1.624861e+00 1.632981e+00 1.640879e+00
#>  [61] 1.648553e+00 1.655999e+00 1.663217e+00 1.670203e+00 1.676956e+00
#>  [66] 1.683473e+00 1.689753e+00 1.695795e+00 1.701595e+00 1.707152e+00
#>  [71] 1.712466e+00 1.717533e+00 1.722353e+00 1.726924e+00 1.731246e+00
#>  [76] 1.735315e+00 1.739133e+00 1.742696e+00 1.746004e+00 1.749057e+00
#>  [81] 1.751853e+00 1.754392e+00 1.756672e+00 1.758693e+00 1.760455e+00
#>  [86] 1.761957e+00 1.763198e+00 1.764178e+00 1.764897e+00 1.765355e+00
#>  [91] 1.765551e+00 1.765486e+00 1.765159e+00 1.764571e+00 1.763721e+00
#>  [96] 1.762610e+00 1.761239e+00 1.759607e+00 1.757715e+00 1.755564e+00
#> [101] 1.753155e+00 1.750487e+00 1.747563e+00 1.744382e+00 1.740946e+00
#> [106] 1.737256e+00 1.733312e+00 1.729116e+00 1.724670e+00 1.719974e+00
#> [111] 1.715030e+00 1.709840e+00 1.704404e+00 1.698725e+00 1.692804e+00
#> [116] 1.686643e+00 1.680244e+00 1.673609e+00 1.666739e+00 1.659637e+00
#> [121] 1.652304e+00 1.644744e+00 1.636958e+00 1.628948e+00 1.620718e+00
#> [126] 1.612268e+00 1.603603e+00 1.594724e+00 1.585634e+00 1.576336e+00
#> [131] 1.566832e+00 1.557125e+00 1.547219e+00 1.537116e+00 1.526818e+00
#> [136] 1.516330e+00 1.505655e+00 1.494794e+00 1.483753e+00 1.472533e+00
#> [141] 1.461138e+00 1.449573e+00 1.437839e+00 1.425941e+00 1.413881e+00
#> [146] 1.401665e+00 1.389294e+00 1.376774e+00 1.364107e+00 1.351297e+00
#> [151] 1.338349e+00 1.325266e+00 1.312051e+00 1.298710e+00 1.285245e+00
#> [156] 1.271661e+00 1.257961e+00 1.244151e+00 1.230233e+00 1.216213e+00
#> [161] 1.202093e+00 1.187879e+00 1.173575e+00 1.159184e+00 1.144712e+00
#> [166] 1.130162e+00 1.115538e+00 1.100846e+00 1.086089e+00 1.071272e+00
#> [171] 1.056399e+00 1.041474e+00 1.026502e+00 1.011488e+00 9.964360e-01
#> [176] 9.813500e-01 9.662349e-01 9.510949e-01 9.359348e-01 9.207589e-01
#> [181] 9.055717e-01 8.903778e-01 8.751817e-01 8.599877e-01 8.448006e-01
#> [186] 8.296247e-01 8.144645e-01 7.993246e-01 7.842094e-01 7.691235e-01
#> [191] 7.540712e-01 7.390570e-01 7.240854e-01 7.091609e-01 6.942878e-01
#> [196] 6.794706e-01 6.647136e-01 6.500212e-01 6.353978e-01 6.208477e-01
#> [201] 6.063752e-01 5.919846e-01 5.776802e-01 5.634662e-01 5.493468e-01
#> [206] 5.353262e-01 5.214086e-01 5.075980e-01 4.938987e-01 4.803145e-01
#> [211] 4.668497e-01 4.535080e-01 4.402936e-01 4.272103e-01 4.142620e-01
#> [216] 4.014525e-01 3.887857e-01 3.762652e-01 3.638948e-01 3.516782e-01
#> [221] 3.396190e-01 3.277207e-01 3.159868e-01 3.044210e-01 2.930265e-01
#> [226] 2.818068e-01 2.707651e-01 2.599048e-01 2.492291e-01 2.387411e-01
#> [231] 2.284440e-01 2.183407e-01 2.084343e-01 1.987278e-01 1.892240e-01
#> [236] 1.799256e-01 1.708356e-01 1.619565e-01 1.532910e-01 1.448417e-01
#> [241] 1.366110e-01 1.286014e-01 1.208154e-01 1.132551e-01 1.059228e-01
#> [246] 9.882073e-02 9.195096e-02 8.531554e-02 7.891642e-02 7.275550e-02
#> [251] 6.683462e-02 6.115551e-02 5.571987e-02 5.052930e-02 4.558535e-02
#> [256] 4.088948e-02 3.644307e-02 3.224746e-02 2.830388e-02 2.461349e-02
#> [261] 2.117740e-02 1.799662e-02 1.507209e-02 1.240469e-02 9.995194e-03
#> [266] 7.844324e-03 5.952717e-03 4.320932e-03 2.949454e-03 1.838688e-03
#> [271] 9.889644e-04 4.005342e-04 7.357205e-05 8.174773e-06 2.043618e-04
#> [276] 6.620749e-04 1.381178e-03 2.361460e-03 3.602627e-03 5.104315e-03
#> [281] 6.866076e-03 8.887389e-03 1.116766e-02 1.370620e-02 1.650227e-02
#> [286] 1.955504e-02 2.286359e-02 2.642696e-02 3.024409e-02 3.431384e-02
#> [291] 3.863501e-02 4.320632e-02 4.802641e-02 5.309385e-02 5.840715e-02
#> [296] 6.396473e-02 6.976495e-02 7.580607e-02 8.208632e-02 8.860383e-02
#> [301] 9.535668e-02 1.023429e-01 1.095603e-01 1.170068e-01 1.246803e-01
#> [306] 1.325784e-01 1.406988e-01 1.490391e-01 1.575969e-01 1.663695e-01
#> [311] 1.753544e-01 1.845489e-01 1.939504e-01 2.035559e-01 2.133627e-01
#> [316] 2.233679e-01 2.335685e-01 2.439614e-01 2.545437e-01 2.653121e-01
#> [321] 2.762635e-01 2.873946e-01 2.987021e-01 3.101827e-01 3.218330e-01
#> [326] 3.336495e-01 3.456287e-01 3.577671e-01 3.700610e-01 3.825069e-01
#> [331] 3.951010e-01 4.078397e-01 4.207190e-01 4.337353e-01 4.468847e-01
#> [336] 4.601632e-01 4.735669e-01 4.870920e-01 5.007342e-01 5.144897e-01
#> [341] 5.283543e-01 5.423239e-01 5.563944e-01 5.705616e-01 5.848214e-01
#> [346] 5.991694e-01 6.136015e-01 6.281133e-01 6.427006e-01 6.573590e-01
#> [351] 6.720843e-01 6.868719e-01 7.017176e-01 7.166170e-01 7.315656e-01
#> [356] 7.465590e-01 7.615928e-01 7.766625e-01 7.917636e-01 8.068918e-01
#> [361] 8.220423e-01 8.372109e-01 8.523930e-01 8.675841e-01 8.827797e-01
#> 
#> [[2]]
#>   [1] 9.814978e-01 9.981019e-01 1.014696e+00 1.031276e+00 1.047835e+00
#>   [6] 1.064371e+00 1.080876e+00 1.097348e+00 1.113780e+00 1.130168e+00
#>  [11] 1.146507e+00 1.162792e+00 1.179018e+00 1.195182e+00 1.211276e+00
#>  [16] 1.227298e+00 1.243242e+00 1.259104e+00 1.274878e+00 1.290561e+00
#>  [21] 1.306147e+00 1.321632e+00 1.337011e+00 1.352280e+00 1.367434e+00
#>  [26] 1.382469e+00 1.397380e+00 1.412163e+00 1.426814e+00 1.441327e+00
#>  [31] 1.455700e+00 1.469927e+00 1.484004e+00 1.497928e+00 1.511693e+00
#>  [36] 1.525297e+00 1.538734e+00 1.552001e+00 1.565095e+00 1.578010e+00
#>  [41] 1.590744e+00 1.603293e+00 1.615652e+00 1.627818e+00 1.639788e+00
#>  [46] 1.651558e+00 1.663125e+00 1.674485e+00 1.685634e+00 1.696570e+00
#>  [51] 1.707289e+00 1.717788e+00 1.728064e+00 1.738113e+00 1.747934e+00
#>  [56] 1.757522e+00 1.766876e+00 1.775992e+00 1.784868e+00 1.793501e+00
#>  [61] 1.801888e+00 1.810027e+00 1.817916e+00 1.825552e+00 1.832933e+00
#>  [66] 1.840056e+00 1.846921e+00 1.853524e+00 1.859863e+00 1.865938e+00
#>  [71] 1.871745e+00 1.877284e+00 1.882553e+00 1.887549e+00 1.892272e+00
#>  [76] 1.896721e+00 1.900893e+00 1.904788e+00 1.908404e+00 1.911741e+00
#>  [81] 1.914797e+00 1.917571e+00 1.920064e+00 1.922273e+00 1.924199e+00
#>  [86] 1.925840e+00 1.927197e+00 1.928268e+00 1.929054e+00 1.929554e+00
#>  [91] 1.929769e+00 1.929697e+00 1.929340e+00 1.928697e+00 1.927768e+00
#>  [96] 1.926554e+00 1.925055e+00 1.923271e+00 1.921204e+00 1.918853e+00
#> [101] 1.916219e+00 1.913304e+00 1.910107e+00 1.906631e+00 1.902875e+00
#> [106] 1.898841e+00 1.894531e+00 1.889945e+00 1.885085e+00 1.879952e+00
#> [111] 1.874549e+00 1.868875e+00 1.862934e+00 1.856727e+00 1.850255e+00
#> [116] 1.843521e+00 1.836527e+00 1.829274e+00 1.821765e+00 1.814003e+00
#> [121] 1.805989e+00 1.797725e+00 1.789215e+00 1.780460e+00 1.771464e+00
#> [126] 1.762229e+00 1.752757e+00 1.743052e+00 1.733117e+00 1.722954e+00
#> [131] 1.712566e+00 1.701956e+00 1.691129e+00 1.680086e+00 1.668831e+00
#> [136] 1.657367e+00 1.645699e+00 1.633828e+00 1.621760e+00 1.609496e+00
#> [141] 1.597042e+00 1.584400e+00 1.571575e+00 1.558570e+00 1.545389e+00
#> [146] 1.532036e+00 1.518515e+00 1.504830e+00 1.490985e+00 1.476984e+00
#> [151] 1.462832e+00 1.448531e+00 1.434088e+00 1.419505e+00 1.404788e+00
#> [156] 1.389941e+00 1.374967e+00 1.359872e+00 1.344660e+00 1.329335e+00
#> [161] 1.313902e+00 1.298366e+00 1.282731e+00 1.267002e+00 1.251184e+00
#> [166] 1.235280e+00 1.219297e+00 1.203238e+00 1.187108e+00 1.170913e+00
#> [171] 1.154656e+00 1.138344e+00 1.121980e+00 1.105569e+00 1.089117e+00
#> [176] 1.072627e+00 1.056106e+00 1.039558e+00 1.022988e+00 1.006401e+00
#> [181] 9.898008e-01 9.731937e-01 9.565841e-01 9.399769e-01 9.233772e-01
#> [186] 9.067897e-01 8.902195e-01 8.736714e-01 8.571503e-01 8.406612e-01
#> [191] 8.242088e-01 8.077982e-01 7.914341e-01 7.751214e-01 7.588649e-01
#> [196] 7.426695e-01 7.265399e-01 7.104810e-01 6.944974e-01 6.785940e-01
#> [201] 6.627754e-01 6.470463e-01 6.314114e-01 6.158753e-01 6.004427e-01
#> [206] 5.851180e-01 5.699059e-01 5.548108e-01 5.398372e-01 5.249896e-01
#> [211] 5.102723e-01 4.956897e-01 4.812462e-01 4.669460e-01 4.527933e-01
#> [216] 4.387924e-01 4.249474e-01 4.112624e-01 3.977414e-01 3.843885e-01
#> [221] 3.712076e-01 3.582026e-01 3.453774e-01 3.327358e-01 3.202815e-01
#> [226] 3.080182e-01 2.959495e-01 2.840791e-01 2.724104e-01 2.609469e-01
#> [231] 2.496920e-01 2.386490e-01 2.278212e-01 2.172119e-01 2.068241e-01
#> [236] 1.966609e-01 1.867253e-01 1.770204e-01 1.675489e-01 1.583137e-01
#> [241] 1.493175e-01 1.405629e-01 1.320526e-01 1.237891e-01 1.157749e-01
#> [246] 1.080122e-01 1.005035e-01 9.325091e-02 8.625660e-02 7.952264e-02
#> [251] 7.305104e-02 6.684371e-02 6.090249e-02 5.522913e-02 4.982534e-02
#> [256] 4.469269e-02 3.983272e-02 3.524686e-02 3.093648e-02 2.690284e-02
#> [261] 2.314715e-02 1.967052e-02 1.647398e-02 1.355847e-02 1.092487e-02
#> [266] 8.573941e-03 6.506391e-03 4.722831e-03 3.223788e-03 2.009708e-03
#> [271] 1.080950e-03 4.377887e-04 8.041514e-05 8.935125e-06 2.233699e-04
#> [276] 7.236558e-04 1.509645e-03 2.581104e-03 3.937715e-03 5.579077e-03
#> [281] 7.504703e-03 9.714023e-03 1.220638e-02 1.498104e-02 1.803718e-02
#> [286] 2.137389e-02 2.499018e-02 2.888499e-02 3.305715e-02 3.750544e-02
#> [291] 4.222853e-02 4.722502e-02 5.249344e-02 5.803222e-02 6.383972e-02
#> [296] 6.991422e-02 7.625392e-02 8.285695e-02 8.972134e-02 9.684506e-02
#> [301] 1.042260e-01 1.118620e-01 1.197507e-01 1.278899e-01 1.362771e-01
#> [306] 1.449098e-01 1.537855e-01 1.629016e-01 1.722553e-01 1.818439e-01
#> [311] 1.916645e-01 2.017142e-01 2.119901e-01 2.224891e-01 2.332080e-01
#> [316] 2.441438e-01 2.552932e-01 2.666528e-01 2.782193e-01 2.899893e-01
#> [321] 3.019593e-01 3.141257e-01 3.264850e-01 3.390334e-01 3.517673e-01
#> [326] 3.646829e-01 3.777763e-01 3.910437e-01 4.044811e-01 4.180846e-01
#> [331] 4.318502e-01 4.457737e-01 4.598510e-01 4.740779e-01 4.884503e-01
#> [336] 5.029639e-01 5.176144e-01 5.323974e-01 5.473085e-01 5.623434e-01
#> [341] 5.774976e-01 5.927665e-01 6.081458e-01 6.236307e-01 6.392168e-01
#> [346] 6.548993e-01 6.706738e-01 6.865354e-01 7.024795e-01 7.185013e-01
#> [351] 7.345962e-01 7.507593e-01 7.669858e-01 7.832710e-01 7.996100e-01
#> [356] 8.159980e-01 8.324301e-01 8.489015e-01 8.654072e-01 8.819424e-01
#> [361] 8.985022e-01 9.150816e-01 9.316758e-01 9.482799e-01 9.648889e-01
#> 
#> [[3]]
#>   [1] 1.109203e+01 1.127967e+01 1.146721e+01 1.165457e+01 1.184172e+01
#>   [6] 1.202858e+01 1.221512e+01 1.240126e+01 1.258696e+01 1.277216e+01
#>  [11] 1.295681e+01 1.314085e+01 1.332423e+01 1.350689e+01 1.368878e+01
#>  [16] 1.386985e+01 1.405003e+01 1.422929e+01 1.440756e+01 1.458479e+01
#>  [21] 1.476093e+01 1.493593e+01 1.510973e+01 1.528228e+01 1.545354e+01
#>  [26] 1.562345e+01 1.579197e+01 1.595903e+01 1.612460e+01 1.628862e+01
#>  [31] 1.645104e+01 1.661182e+01 1.677091e+01 1.692826e+01 1.708383e+01
#>  [36] 1.723756e+01 1.738942e+01 1.753936e+01 1.768733e+01 1.783329e+01
#>  [41] 1.797720e+01 1.811901e+01 1.825868e+01 1.839618e+01 1.853145e+01
#>  [46] 1.866446e+01 1.879518e+01 1.892356e+01 1.904956e+01 1.917314e+01
#>  [51] 1.929428e+01 1.941293e+01 1.952906e+01 1.964263e+01 1.975362e+01
#>  [56] 1.986198e+01 1.996768e+01 2.007071e+01 2.017101e+01 2.026857e+01
#>  [61] 2.036336e+01 2.045534e+01 2.054449e+01 2.063078e+01 2.071420e+01
#>  [66] 2.079470e+01 2.087228e+01 2.094690e+01 2.101855e+01 2.108719e+01
#>  [71] 2.115283e+01 2.121542e+01 2.127496e+01 2.133143e+01 2.138480e+01
#>  [76] 2.143507e+01 2.148222e+01 2.152624e+01 2.156711e+01 2.160481e+01
#>  [81] 2.163935e+01 2.167071e+01 2.169888e+01 2.172384e+01 2.174561e+01
#>  [86] 2.176415e+01 2.177949e+01 2.179159e+01 2.180048e+01 2.180613e+01
#>  [91] 2.180855e+01 2.180775e+01 2.180371e+01 2.179644e+01 2.178594e+01
#>  [96] 2.177222e+01 2.175528e+01 2.173513e+01 2.171176e+01 2.168519e+01
#> [101] 2.165543e+01 2.162248e+01 2.158636e+01 2.154707e+01 2.150462e+01
#> [106] 2.145904e+01 2.141033e+01 2.135850e+01 2.130358e+01 2.124557e+01
#> [111] 2.118450e+01 2.112039e+01 2.105325e+01 2.098310e+01 2.090996e+01
#> [116] 2.083386e+01 2.075482e+01 2.067285e+01 2.058800e+01 2.050027e+01
#> [121] 2.040970e+01 2.031631e+01 2.022014e+01 2.012120e+01 2.001953e+01
#> [126] 1.991516e+01 1.980813e+01 1.969845e+01 1.958617e+01 1.947131e+01
#> [131] 1.935392e+01 1.923402e+01 1.911165e+01 1.898686e+01 1.885966e+01
#> [136] 1.873011e+01 1.859824e+01 1.846409e+01 1.832770e+01 1.818911e+01
#> [141] 1.804837e+01 1.790550e+01 1.776056e+01 1.761359e+01 1.746463e+01
#> [146] 1.731373e+01 1.716093e+01 1.700627e+01 1.684981e+01 1.669158e+01
#> [151] 1.653164e+01 1.637003e+01 1.620680e+01 1.604201e+01 1.587568e+01
#> [156] 1.570789e+01 1.553867e+01 1.536808e+01 1.519616e+01 1.502298e+01
#> [161] 1.484857e+01 1.467300e+01 1.449631e+01 1.431855e+01 1.413978e+01
#> [166] 1.396005e+01 1.377942e+01 1.359794e+01 1.341566e+01 1.323263e+01
#> [171] 1.304891e+01 1.286456e+01 1.267963e+01 1.249417e+01 1.230824e+01
#> [176] 1.212189e+01 1.193519e+01 1.174818e+01 1.156091e+01 1.137346e+01
#> [181] 1.118586e+01 1.099818e+01 1.081047e+01 1.062279e+01 1.043520e+01
#> [186] 1.024774e+01 1.006048e+01 9.873468e+00 9.686761e+00 9.500415e+00
#> [191] 9.314485e+00 9.129027e+00 8.944094e+00 8.759742e+00 8.576026e+00
#> [196] 8.392999e+00 8.210717e+00 8.029233e+00 7.848600e+00 7.668874e+00
#> [201] 7.490106e+00 7.312350e+00 7.135658e+00 6.960083e+00 6.785676e+00
#> [206] 6.612490e+00 6.440576e+00 6.269985e+00 6.100766e+00 5.932972e+00
#> [211] 5.766650e+00 5.601851e+00 5.438623e+00 5.277014e+00 5.117073e+00
#> [216] 4.958847e+00 4.802383e+00 4.647727e+00 4.494925e+00 4.344022e+00
#> [221] 4.195063e+00 4.048092e+00 3.903153e+00 3.760288e+00 3.619540e+00
#> [226] 3.480951e+00 3.344562e+00 3.210413e+00 3.078543e+00 2.948993e+00
#> [231] 2.821800e+00 2.697002e+00 2.574636e+00 2.454738e+00 2.337344e+00
#> [236] 2.222489e+00 2.110206e+00 2.000529e+00 1.893491e+00 1.789122e+00
#> [241] 1.687455e+00 1.588519e+00 1.492343e+00 1.398956e+00 1.308386e+00
#> [246] 1.220660e+00 1.135802e+00 1.053840e+00 9.747964e-01 8.986951e-01
#> [251] 8.255588e-01 7.554089e-01 6.882665e-01 6.241512e-01 5.630822e-01
#> [256] 5.050776e-01 4.501544e-01 3.983291e-01 3.496169e-01 3.040323e-01
#> [261] 2.615888e-01 2.222990e-01 1.861745e-01 1.532260e-01 1.234633e-01
#> [266] 9.689516e-02 7.352952e-02 5.337329e-02 3.643243e-02 2.271196e-02
#> [271] 1.221595e-02 4.947504e-03 9.087813e-04 1.009769e-04 2.524330e-03
#> [276] 8.178123e-03 1.706068e-02 2.916937e-02 4.450060e-02 6.304984e-02
#> [281] 8.481158e-02 1.097794e-01 1.379458e-01 1.693026e-01 2.038404e-01
#> [286] 2.415489e-01 2.824171e-01 3.264328e-01 3.735829e-01 4.238536e-01
#> [291] 4.772298e-01 5.336958e-01 5.932348e-01 6.558292e-01 7.214605e-01
#> [296] 7.901092e-01 8.617550e-01 9.363765e-01 1.013952e+00 1.094458e+00
#> [301] 1.177871e+00 1.264166e+00 1.353317e+00 1.445299e+00 1.540084e+00
#> [306] 1.637644e+00 1.737949e+00 1.840971e+00 1.946678e+00 2.055040e+00
#> [311] 2.166024e+00 2.279597e+00 2.395726e+00 2.514376e+00 2.635513e+00
#> [316] 2.759099e+00 2.885099e+00 3.013476e+00 3.144191e+00 3.277205e+00
#> [321] 3.412479e+00 3.549973e+00 3.689647e+00 3.831458e+00 3.975366e+00
#> [326] 4.121326e+00 4.269296e+00 4.419233e+00 4.571091e+00 4.724826e+00
#> [331] 4.880392e+00 5.037743e+00 5.196832e+00 5.357613e+00 5.520037e+00
#> [336] 5.684057e+00 5.849624e+00 6.016688e+00 6.185201e+00 6.355112e+00
#> [341] 6.526371e+00 6.698927e+00 6.872730e+00 7.047727e+00 7.223867e+00
#> [346] 7.401098e+00 7.579367e+00 7.758621e+00 7.938807e+00 8.119872e+00
#> [351] 8.301762e+00 8.484423e+00 8.667801e+00 8.851842e+00 9.036491e+00
#> [356] 9.221694e+00 9.407395e+00 9.593540e+00 9.780073e+00 9.966939e+00
#> [361] 1.015408e+01 1.034145e+01 1.052898e+01 1.071663e+01 1.090433e+01
#>